localhost altered

+++ src/Pages/Inventory/InventoryItem/InventoryItem.js
@@ -2,17 +2,20 @@ import React from "react";
 
 const InventoryItem = ({ item }) => {
   const { _id, img, name, supplierName, price, quantity } = item;
-  const handleDelete = () => {
-    fetch(`http://localhost:5000/inventory/${_id}`, {
-      method: "DELETE",
-      headers: {
-        "content-type": "application/json",
-      },
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-      });
+  const handleItemDelete = (id) => {
+    const confirmed = window.confirm("Sure to delete?");
+    if (confirmed) {
+      fetch(`https://blooming-wave-56097.herokuapp.com/inventory/${id}`, {
+        method: "DELETE",
+        headers: {
+          "content-type": "application/json",
+        },
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          console.log(data);
+        });
+    }
   };
   return (
     <tr className="">
@@ -25,7 +28,7 @@ const InventoryItem = ({ item }) => {
       <td>{price}</td>
       <td>{quantity}</td>
       <td>
-        <button className="btn btn-dark">
+        <button onClick={() => handleItemDelete(_id)} className="btn btn-dark">
           <img
             style={{ width: "32px", height: "32px" }}
             src="delete.png"
